% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rschrodinger-package.R
\docType{package}
\name{rschrodinger-package}
\alias{rschrodinger}
\alias{rschrodinger-package}
\title{rschrodinger: Fast Data Bridge Between Schrodinger and R}
\description{
Read Maestro (.mae/.maegz) files natively in R without requiring a Schrodinger installation. Optionally bridge to Schrodinger's Python API for protein-ligand interaction analysis and job monitoring via sandboxed subprocess calls.
}
\author{
\strong{Maintainer}: Jixing Liu \email{jixing.liu@example.com} [copyright holder]

}
\keyword{internal}
